pyproject.toml
[project]
name = "ddns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "psutil>=7.0.0",
    "pydantic-settings>=2.9.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.13.0",
]